Use parent config class for common config properties
