+++ src/main/java/com/microsoft/azure/cosmosdb/kafka/connect/source/CosmosDBSourceConfig.java
@@ -1,6 +1,5 @@
 package com.microsoft.azure.cosmosdb.kafka.connect.source;
 
-import org.apache.kafka.common.config.AbstractConfig;
 import org.apache.kafka.common.config.ConfigDef;
 import org.apache.kafka.common.config.ConfigDef.Importance;
 import org.apache.kafka.common.config.ConfigDef.Type;
@@ -9,32 +8,14 @@ import org.apache.kafka.common.config.ConfigDef.Width;
 import java.util.Arrays;
 import java.util.Map;
 
-import com.microsoft.azure.cosmosdb.kafka.connect.TopicContainerMap;
+import com.microsoft.azure.cosmosdb.kafka.connect.CosmosDBConfig;
 
 /**
  * Contains settings for the Kafka ComsosDB Source Connector
  */
 
 @SuppressWarnings ({"squid:S1854", "squid:S2160"})  // suppress unneeded int *groupOrder variables, equals method
-public class CosmosDBSourceConfig extends AbstractConfig {
-
-    static final String COSMOS_CONN_ENDPOINT_CONF = "connect.cosmosdb.connection.endpoint";
-    private static final String COSMOS_CONN_ENDPOINT_DOC = "Cosmos DB endpoint URL.";
-    private static final String COSMOS_CONN_ENDPOINT_DISPLAY = "Cosmos DB Endpoint";
-
-    static final String COSMOS_CONN_KEY_CONF = "connect.cosmosdb.master.key";
-    private static final String COSMOS_CONN_KEY_DOC = "Cosmos DB connection master (primary) key.";
-    private static final String COSMOS_CONN_KEY_DISPLAY = "Cosmos DB Connection Key";
-
-    static final String COSMOS_DATABASE_NAME_CONF = "connect.cosmosdb.databasename";
-    private static final String COSMOS_DATABASE_NAME_DOC = "Cosmos DB target database to read records from.";
-    private static final String COSMOS_DATABASE_NAME_DISPLAY = "Cosmos DB Database name";
-
-    static final String COSMOS_CONTAINER_TOPIC_MAP_CONF = "connect.cosmosdb.containers.topicmap";
-    private static final String COSMOS_CONTAINER_TOPIC_MAP_DOC = 
-        "A comma delimited list of Kafka topics mapped to Cosmos DB containers.\n" +
-        "For example: topic1#con1,topic2#con2.";
-    private static final String COSMOS_CONTAINER_TOPIC_MAP_DISPLAY = "Topic-Container map";
+public class CosmosDBSourceConfig extends CosmosDBConfig {
 
     static final String COSMOS_SOURCE_TASK_TIMEOUT_CONF = "connect.cosmosdb.task.timeout";
     private static final Long COSMOS_SOURCE_TASK_TIMEOUT_DEFAULT = 5000L;
@@ -90,12 +71,6 @@ public class CosmosDBSourceConfig extends AbstractConfig {
         "Whether the change feed should start from the beginning.";
     private static final String COSMOS_CF_START_FROM_BEGINNING_DISPLAY = "Cosmos Change Feed start from beginning";
 
-    public static final String COSMOS_CLIENT_USER_AGENT_SUFFIX = "APN/1.0 Microsoft/1.0 KafkaConnect/";
-
-    private String connEndpoint;
-    private String connKey;
-    private String databaseName;
-    private TopicContainerMap topicContainerMap = TopicContainerMap.empty();
     private Long timeout;
     private Long bufferSize;
     private Long batchSize;
@@ -112,12 +87,8 @@ public class CosmosDBSourceConfig extends AbstractConfig {
     }
 
     public CosmosDBSourceConfig(Map<String, String> parsedConfig) {
-        this(getConfig(), parsedConfig);
+        super(getConfig(), parsedConfig);
 
-        connEndpoint = this.getString(COSMOS_CONN_ENDPOINT_CONF);
-        connKey = this.getPassword(COSMOS_CONN_KEY_CONF).value();
-        databaseName = this.getString(COSMOS_DATABASE_NAME_CONF);
-        topicContainerMap = TopicContainerMap.deserialize(this.getString(COSMOS_CONTAINER_TOPIC_MAP_CONF));
         timeout = this.getLong(COSMOS_SOURCE_TASK_TIMEOUT_CONF);
         bufferSize = this.getLong(COSMOS_SOURCE_TASK_BUFFER_SIZE_CONF);
         batchSize = this.getLong(COSMOS_SOURCE_TASK_BATCH_SIZE_CONF);
@@ -131,43 +102,13 @@ public class CosmosDBSourceConfig extends AbstractConfig {
     }
 
     public static ConfigDef getConfig() {
-        ConfigDef result = new ConfigDef();
+        ConfigDef result = CosmosDBConfig.getConfig();
         
-        defineConnectionConfigs(result);
         defineTaskConfigs(result);
         defineDatabaseConfigs(result);
         defineMessageConfigs(result);
 
         return result;
-    }
-
-    private static void defineConnectionConfigs(ConfigDef result) {
-        final String connectionGroupName = "Connection";
-        int connectionGroupOrder = 0;
-        
-        result
-            .define(
-                COSMOS_CONN_ENDPOINT_CONF,
-                Type.STRING,
-                ConfigDef.NO_DEFAULT_VALUE,
-                Importance.HIGH,
-                COSMOS_CONN_ENDPOINT_DOC,
-                connectionGroupName,
-                connectionGroupOrder++,
-                Width.LONG,
-                COSMOS_CONN_ENDPOINT_DISPLAY
-            )
-            .define(
-                COSMOS_CONN_KEY_CONF,
-                Type.PASSWORD,
-                ConfigDef.NO_DEFAULT_VALUE,
-                Importance.HIGH,
-                COSMOS_CONN_KEY_DOC,
-                connectionGroupName,
-                connectionGroupOrder++,
-                Width.LONG,
-                COSMOS_CONN_KEY_DISPLAY
-            );
     }
 
     private static void defineTaskConfigs(ConfigDef result) {
@@ -224,31 +165,9 @@ public class CosmosDBSourceConfig extends AbstractConfig {
     
     private static void defineDatabaseConfigs(ConfigDef result) {
         final String databaseGroupName = "Database";
-        int databaseGroupOrder = 0;
+        int databaseGroupOrder = CosmosDBConfig.COSMOS_DATABASE_GROUP_ORDER;
         
         result
-            .define(
-                COSMOS_DATABASE_NAME_CONF,
-                Type.STRING,
-                ConfigDef.NO_DEFAULT_VALUE,
-                Importance.HIGH,
-                COSMOS_DATABASE_NAME_DOC,
-                databaseGroupName,
-                databaseGroupOrder++,
-                Width.MEDIUM,
-                COSMOS_DATABASE_NAME_DISPLAY
-            )
-            .define(
-                COSMOS_CONTAINER_TOPIC_MAP_CONF,
-                Type.STRING,
-                ConfigDef.NO_DEFAULT_VALUE,
-                Importance.HIGH,
-                COSMOS_CONTAINER_TOPIC_MAP_DOC,
-                databaseGroupName,
-                databaseGroupOrder++,
-                Width.MEDIUM,
-                COSMOS_CONTAINER_TOPIC_MAP_DISPLAY
-            )
             .define(
                 COSMOS_CONTAINERS_LIST_CONF,
                 Type.STRING,
@@ -256,7 +175,7 @@ public class CosmosDBSourceConfig extends AbstractConfig {
                 Importance.HIGH,
                 COSMOS_CONTAINERS_LIST_DOC,
                 databaseGroupName,
-                databaseGroupOrder++,
+                ++databaseGroupOrder,
                 Width.MEDIUM,
                 COSMOS_CONTAINERS_LIST_DISPLAY
             )
@@ -267,7 +186,7 @@ public class CosmosDBSourceConfig extends AbstractConfig {
                 Importance.HIGH,
                 COSMOS_CF_START_FROM_BEGINNING_DOC,
                 databaseGroupName,
-                databaseGroupOrder++,
+                ++databaseGroupOrder,
                 Width.MEDIUM,
                 COSMOS_CF_START_FROM_BEGINNING_DISPLAY
             )
@@ -278,7 +197,7 @@ public class CosmosDBSourceConfig extends AbstractConfig {
                 Importance.MEDIUM,
                 COSMOS_ASSIGNED_CONTAINER_DOC,
                 databaseGroupName,
-                databaseGroupOrder++,
+                ++databaseGroupOrder,
                 Width.MEDIUM,
                 COSMOS_ASSIGNED_CONTAINER_DISPLAY
             )
@@ -289,7 +208,7 @@ public class CosmosDBSourceConfig extends AbstractConfig {
                 Importance.MEDIUM,
                 COSMOS_WORKER_NAME_DOC,
                 databaseGroupName,
-                databaseGroupOrder++,
+                ++databaseGroupOrder,
                 Width.MEDIUM,
                 COSMOS_WORKER_NAME_DISPLAY
             );
@@ -323,22 +242,6 @@ public class CosmosDBSourceConfig extends AbstractConfig {
                 COSMOS_MESSAGE_KEY_FIELD_DISPLAY,
                 Arrays.asList(COSMOS_MESSAGE_KEY_ENABLED_CONF)
             );
-    }
-
-    public String getConnEndpoint() {
-        return this.connEndpoint;
-    }
-
-    public String getConnKey() {
-        return this.connKey;
-    }
-
-    public String getDatabaseName() {
-        return this.databaseName;
-    }
-
-    public TopicContainerMap getTopicContainerMap() {
-        return this.topicContainerMap;
     }
 
     public String getContainerList() {
